pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Rodrigo Sumioshi <43658223+sumioshi@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
factory-boy = "^3.3.0"
django = "^5.1"
djangorestframework = "^3.14.0"  # Adicione esta linha
pillow = "^9.0.0"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
